feat: implement context compaction configuration and utility functions
