+++ crates/forge_domain/src/agent.rs
@@ -4,13 +4,13 @@ use derive_more::derive::Display;
 use derive_setters::Setters;
 use merge::Merge;
 use serde::{Deserialize, Serialize};
-use tracing::debug;
 
+use crate::compact::Compact;
 use crate::merge::Key;
 use crate::temperature::Temperature;
 use crate::template::Template;
 use crate::{
-    Context, Error, Event, EventContext, ModelId, Result, Role, SystemContext, ToolDefinition,
+    Context, Error, Event, EventContext, ModelId, Result, SystemContext, ToolDefinition,
     ToolName, TopK, TopP,
 };
 
@@ -30,116 +30,6 @@ impl AgentId {
     }
 }
 
-/// Configuration for automatic context compaction
-#[derive(Debug, Clone, Serialize, Deserialize, Merge, Setters)]
-#[setters(strip_option, into)]
-pub struct Compact {
-    /// Number of most recent messages to preserve during compaction
-    /// These messages won't be considered for summarization
-    #[merge(strategy = crate::merge::std::overwrite)]
-    pub retention_window: usize,
-    /// Maximum number of tokens to keep after compaction
-    #[merge(strategy = crate::merge::option)]
-    pub max_tokens: Option<usize>,
-
-    /// Maximum number of tokens before triggering compaction
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[merge(strategy = crate::merge::option)]
-    pub token_threshold: Option<u64>,
-
-    /// Maximum number of conversation turns before triggering compaction
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[merge(strategy = crate::merge::option)]
-    pub turn_threshold: Option<usize>,
-
-    /// Maximum number of messages before triggering compaction
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[merge(strategy = crate::merge::option)]
-    pub message_threshold: Option<usize>,
-
-    /// Optional custom prompt template to use during compaction
-    #[serde(skip_serializing_if = "Option::is_none")]
-    #[merge(strategy = crate::merge::option)]
-    pub prompt: Option<String>,
-
-    /// Model ID to use for compaction, useful when compacting with a
-    /// cheaper/faster model
-    #[merge(strategy = crate::merge::std::overwrite)]
-    pub model: ModelId,
-    /// Optional tag name to extract content from when summarizing (e.g.,
-    /// "summary")
-    #[merge(strategy = crate::merge::std::overwrite)]
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub summary_tag: Option<SummaryTag>,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-#[serde(transparent)]
-pub struct SummaryTag(String);
-
-impl Default for SummaryTag {
-    fn default() -> Self {
-        SummaryTag("forge_context_summary".to_string())
-    }
-}
-
-impl SummaryTag {
-    pub fn as_str(&self) -> &str {
-        self.0.as_str()
-    }
-}
-
-impl Compact {
-    /// Creates a new compaction configuration with the specified maximum token
-    /// limit
-    pub fn new(model: ModelId) -> Self {
-        Self {
-            max_tokens: None,
-            token_threshold: None,
-            turn_threshold: None,
-            message_threshold: None,
-            prompt: None,
-            summary_tag: None,
-            model,
-            retention_window: 0,
-        }
-    }
-
-    /// Determines if compaction should be triggered based on the current
-    /// context
-    pub fn should_compact(&self, context: &Context, token_count: u64) -> bool {
-        // Check if any of the thresholds have been exceeded
-        if let Some(token_threshold) = self.token_threshold {
-            debug!(tokens = ?token_count, "Token count");
-            // use provided prompt_tokens if available, otherwise estimate token count
-            if token_count >= token_threshold {
-                return true;
-            }
-        }
-
-        if let Some(turn_threshold) = self.turn_threshold {
-            if context
-                .messages
-                .iter()
-                .filter(|message| message.has_role(Role::User))
-                .count()
-                >= turn_threshold
-            {
-                return true;
-            }
-        }
-
-        if let Some(message_threshold) = self.message_threshold {
-            // Count messages directly from context
-            let msg_count = context.messages.len();
-            if msg_count >= message_threshold {
-                return true;
-            }
-        }
-
-        false
-    }
-}
 #[derive(Debug, Clone, Serialize, Deserialize, Merge, Setters)]
 #[setters(strip_option, into)]
 pub struct Agent {
