+++ crates/forge_domain/src/lib.rs
@@ -3,6 +3,7 @@ mod api;
 mod attachment;
 mod chat_request;
 mod chat_response;
+mod compact;
 mod compaction_result;
 mod conversation_html;
 mod update;
@@ -48,6 +49,7 @@ pub use api::*;
 pub use attachment::*;
 pub use chat_request::*;
 pub use chat_response::*;
+pub use compact::*;
 pub use compaction_result::*;
 pub use context::*;
 pub use conversation::*;
